+++ src/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the Prog8 Language Support extension will be documented in this file.
 
+## [1.6.0] - 2026-02-06
+
+- New project system!
+- F5 now compiles the current code file if you don't have a project file in the same folder as the current file.
+- Use the command `Prog8: Initialize Project` to generate a project file in the current directory.
+
 ## [1.5.1] - 2026-02-06
 
 - Fixed symbol detection on parameters of subs declared in the builtin modules.